[rando] Refactor randomizer code
The main theory behind the refactor is to parse all the data tables
at the beginning of the randomization, keeping a single source of
truth for this data.  The randomizer then manipulates the parsed
data to shuffle the seed.  At the end the parsed data is
re-encoded and written back out to the ROM.

This allows us to more easily patch conditional expressions `0x0b`
which this patch also implements.